requirements.txt
boto==2.38.0
decorator==4.0.2
Django==1.8.4
django-ckeditor==4.5.1
django-extensions==1.5.6
django-redis-cache==1.6.0
django-storages==1.1.8
hiredis==0.2.0
ipdb==0.8.1
ipython==4.0.0
ipython-genutils==0.1.0
path.py==8.1
pexpect==3.3
pickleshare==0.5
Pillow==2.9.0
psycopg2==2.6.1
redis==2.10.3
simplegeneric==0.8.1
six==1.9.0
sorl-thumbnail==12.3
traitlets==4.0.0
wheel==0.24.0
